requirements/app.txt
galaxy.plugin.api==0.66
certifi==2019.3.9
